+++ src/index.ts
@@ -3,7 +3,12 @@ export {
   createMatchPath,
   matchFromAbsolutePaths,
   MatchPath
-} from "./match-path";
+} from "./match-path-sync";
+export {
+  createMatchPathAsync,
+  matchFromAbsolutePathsAsync,
+  MatchPathAsync
+} from "./match-path-async";
 export { register } from "./register";
 export {
   loadConfig,
@@ -11,3 +16,9 @@ export {
   ConfigLoaderSuccessResult,
   ConfigLoaderFailResult
 } from "./config-loader";
+export {
+  ReadJsonSync,
+  ReadJsonAsync,
+  FileExistsSync,
+  FileExistsAsync
+} from "./filesystem";